backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangChain and AI
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-core==0.1.10

# Vector Stores
pinecone-client==3.0.0
chromadb==0.4.18

# Embeddings
sentence-transformers==2.2.2

# Environment and Utils
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Optional (if you need Groq later)
groq==0.4.1
langchain-groq==0.0.1# Up
date Management Dependencies
watchdog==3.0.0
beautifulsoup4==4.12.2
requests==2.31.0